[linux-packaging-snippets] arch: introduce host arch autodetection
This change adds the host arch auto detection capabily
to the kernel-snippet.mk.

After this implementation, the host architecture where the kernel
is going to be compiled, is auto detected.

This avoids the requirement to configure the variable 'DEB_BUILD_ON'
in the kernel-info.mk, which is specilly useful for scenarios
where the host arch changes frequently.

To enable this feature just remove (or comment) the 'DEB_BUILD_ON'
variable from the kernel-info.mk.

Anyway, for backward reasons, the logic located in the kernel-snippet.mk
has no preference if the variable is still defined in the kernel-info.mk.

Also the feature is documented in the kernel-info.mk.example.